pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "lxml>=5.3.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "validators>=0.34.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]
